+++ color/README.md
@@ -1,0 +1,3 @@
+# Кастомный код для работы с цветом
+
+Это не такая эпичная и абстрактная фигня, как с модулями `combo` и `lang_shift`, так что здесь не буду писать огромную ридмишку как это использовать. Читайте код, если вам нужно.
