+++ combo/keycodes.h
@@ -598,3 +598,4 @@
 #if COMBO_KEYS_COUNT > 199
   CMB_199,
 #endif
+
