+++ lang_shift/README.md
@@ -31,6 +31,45 @@ LA_CTSH, /* Задаёт переключение языка на Ctrl + Shift. 
 LA_WISP, /* Задаёт переключение языка на Win + Space. */
 ```
 
+## Задание SAFE_RANGE
+
+Для пользовательских кейкодов существует такое понятие как `SAFE_RANGE`, обычно он используется для того чтобы после него помещать свои кейкоды, которые делают особые действия:
+
+```c
+enum custom_keycodes {
+  MY_KEY1 = SAFE_RANGE,
+  MY_KEY2,
+  // ...
+}
+```
+
+Таким образом гарантируется что пользовательские кейкоды не пересекутся с системными кейкодами. Обычно для этого используется `SAFE_RANGE`.
+
+Но иногда случается так, что клавиатура добавляет своих кейкодов, и создаёт новую переменную SAFE_RANGE, которую называет по-другому, нужно за этим следить. Посмотрите, нету ли у вас такой особой переменной. Например, в Moonlander эта переменная называется `ML_SAFE_RANGE`.
+
+После того как вы найдёте имя этой переменной, нужно написать:
+
+```c
+#define CUSTOM_SAFE_RANGE <ваша переменная>
+```
+
+Пример для Moonlander:
+```c
+#define CUSTOM_SAFE_RANGE ML_SAFE_RANGE
+```
+
+Затем при использовании своих кейкодов необходимо указывать именно `CUSTOM_SAFE_RANGE`, потому что данный модуль использует необходимое количество кейкодов из пользовательских кейкодов и переопределяет эту переменную. Пример:
+
+```c
+enum custom_keycodes {
+  KEYCODES_START = CUSTOM_SAFE_RANGE,
+
+  // English specific keys
+  EN_LTEQ, // <=
+  // ...
+};
+```
+
 ## Подключить код
 
 В своём файле `keymap.c` в самом верху подключаем файл `lang_shift/code.c`:
@@ -105,20 +144,95 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
 **Минусы:**
 * Невозможно посылать кейкоды, не существующие ни в русской, ни в английской раскладке, например `« » —`.
 * Необходимость компилировать свою раскладку из кода, и вообще притрагиваться к коду. Некоторые люди предпочитают работать исключительно в графических конфигураторах.
+* Ввод английских символов при включённой русской раскладке может тормозить из-за тормознутого метода переключения раскладки. Из-за этого же иногда некоторые клавиши могут прожиматься не на том языке.
 
 # Использование в коде
+
+Всё что вы можете использовать в коде можно посмотреть в файле `include.h`.
 
 ## Обычная клавиша на шифтовом слое
 
 Если вы хотите использовать кастомный кейкод на шифтовом слое, но чтобы он был без зажатого шифта, то нужно в коде этого кейкода перед его активацией написать: `shift_activate(0);`. Данный код вырубает шифт, но таким образом что шифт потом сам включится через 200мс, или если он понадобится другой клавише.
 
-# TODO
+Пример использования:
+```c
+bool process_record_user(uint16_t keycode, keyrecord_t *record) {
+	switch (keycode) {
+	  case CT_Y:
+	    if (record->event.pressed) {
+	      shift_activate(0);
+	      register_code(KC_LCTRL);
+	      register_code(KC_Y);
+	      unregister_code(KC_Y);
+	      unregister_code(KC_LCTRL);
+	    }
+	    return false;
+	    break;
+	}
+	return true;
+}
+```
 
-* [ ] Добавить функции
-	* [ ] Чтобы из любого места можно было нажать языково-специфичную клавишу
-	* [ ] Чтобы переключиться на нужный языковой слой
-	* [ ] Чтобы к следующей клавише применить автошифт
-	* [ ] 
-	* [ ] 
-	* [ ] 
-	* [ ] 
+## Англоязычный хоткей
+
+Если вы хотите чтобы перед нажатием вашего хоткея был включён английский язык, то можно перед его нажатием вызывать функцию `lang_activate(0);`. Это будет работать аналогично верхнему пункту. 
+
+Пример использования:
+```c
+bool process_record_user(uint16_t keycode, keyrecord_t *record) {
+	switch (keycode) {
+	  case CT_D:
+	    if (record->event.pressed) {
+	      lang_activate(0);
+	      register_code(KC_LCTRL);
+	      register_code(KC_D);
+	      unregister_code(KC_D);
+	      unregister_code(KC_LCTRL);
+	    }
+	    return false;
+	    break;
+	}
+	return true;
+}
+```
+
+## Набрать символ на нужном языке
+
+Если вы хотите набрать символ из определённого языка, при этом не зная какой включен сейчас, можно использовать функцию `lang_shift_tap_key`. Аналогично можно использовать функцию `lang_shift_press_key` если вам нужно этот символ зажать или отжать.
+
+Пример использования:
+```c
+bool process_record_user(uint16_t keycode, keyrecord_t *record) {
+	switch (keycode) {
+	  case EN_LTEQ:
+	    if (record->event.pressed) {
+	      lang_shift_tap_key(EN_LT);
+	      lang_shift_tap_key(AG_EQL);
+	    }
+	    return false;
+	    break;
+	}
+	return true;
+}
+```
+
+## Применить шифт к следующей клавише
+
+Например, вы хотите создать клавишу `AG_QSSP`, которая нажимает: <вопросительный знак> + <пробел> + <применить шифт к следующей клавише>, то вы можете использовать функцию `shift_once_use_to_next_key`, где в параметрах нужно указать на какой слой нужно переключить следующую клавишу. Если это русский язык, то нужно указать 3; если английский - 0; если вы хотите чтобы оно автоматически подстраивалось под ваш язык, то можно вызвать её с другой функцией: `shift_once_use_to_next_key(lang_get_shift_layer_number())`. В итоге код получится такой:
+
+```c
+bool process_record_user(uint16_t keycode, keyrecord_t *record) {
+	switch (keycode) {
+	  case AG_QSSP:
+	    if (record->event.pressed) {
+	      lang_shift_tap_key(AG_QUES);
+	      register_code(KC_SPC);
+	      unregister_code(KC_SPC);
+	      shift_once_use_to_next_key(lang_get_shift_layer_number());
+	    }
+	    return false;
+	    break;
+	}
+	return true;
+}
+```
