--- lang_shift/keycodes/lang.h
@@ -1,226 +1,0 @@
-/* Английские символы */
-EN_GRV,  /* ` */
-EN_TILD, /* ~ */
-EN_1,    /* 1 */
-EN_EXCL, /* ! */
-EN_2,    /* 2 */
-EN_AT,   /* @ */
-EN_3,    /* 3 */
-EN_HASH, /* # */
-EN_4,    /* 4 */
-EN_DLR,  /* $ */
-EN_5,    /* 5 */
-EN_PERC, /* % */
-EN_6,    /* 6 */
-EN_CIRC, /* ^ */
-EN_7,    /* 7 */
-EN_AMPR, /* & */
-EN_8,    /* 8 */
-EN_ASTR, /* * */
-EN_9,    /* 9 */
-EN_LPRN, /* ( */
-EN_0,    /* 0 */
-EN_RPRN, /* ) */
-EN_MINS, /* - */
-EN_UNDS, /* _ */
-EN_EQL,  /* = */
-EN_PLUS, /* + */
-
-EN_Q,    /* q */
-EN_S_Q,  /* Q */
-EN_W,    /* w */
-EN_S_W,  /* W */
-EN_E,    /* e */
-EN_S_E,  /* E */
-EN_R,    /* r */
-EN_S_R,  /* R */
-EN_T,    /* t */
-EN_S_T,  /* T */
-EN_Y,    /* y */
-EN_S_Y,  /* Y */
-EN_U,    /* u */
-EN_S_U,  /* U */
-EN_I,    /* i */
-EN_S_I,  /* I */
-EN_O,    /* o */
-EN_S_O,  /* O */
-EN_P,    /* p */
-EN_S_P,  /* P */
-EN_LBRC, /* [ */
-EN_LCBR, /* { */
-EN_RBRC, /* ] */
-EN_RCBR, /* } */
-
-EN_A,    /* a */
-EN_S_A,  /* A */
-EN_S,    /* s */
-EN_S_S,  /* S */
-EN_D,    /* d */
-EN_S_D,  /* D */
-EN_F,    /* f */
-EN_S_F,  /* F */
-EN_G,    /* g */
-EN_S_G,  /* G */
-EN_H,    /* h */
-EN_S_H,  /* H */
-EN_J,    /* j */
-EN_S_J,  /* J */
-EN_K,    /* k */
-EN_S_K,  /* K */
-EN_L,    /* l */
-EN_S_L,  /* L */
-EN_SCLN, /* ; */
-EN_COLN, /* : */
-EN_QUOT, /* ' */
-EN_DQUO, /* " */
-EN_BSLS, /* \ */
-EN_PIPE, /* | */
-
-EN_Z,    /* z */
-EN_S_Z,  /* Z */
-EN_X,    /* x */
-EN_S_X,  /* X */
-EN_C,    /* c */
-EN_S_C,  /* C */
-EN_V,    /* v */
-EN_S_V,  /* V */
-EN_B,    /* b */
-EN_S_B,  /* B */
-EN_N,    /* n */
-EN_S_N,  /* N */
-EN_M,    /* m */
-EN_S_M,  /* M */
-EN_COMM, /* , */
-EN_LT,   /* < */
-EN_DOT,  /* . */
-EN_GT,   /* > */
-EN_SLSH, /* / */
-EN_QUES, /* ? */
-/* -------------------------------------------------------------------- */
-/* Русские символы */
-RU_JO,   /* ё */
-RU_S_JO, /* Ё */
-RU_1,    /* 1 */
-RU_EXCL, /* ! */
-RU_2,    /* 2 */
-RU_DQUO, /* " */
-RU_3,    /* 3 */
-RU_NUME, /* № */
-RU_4,    /* 4 */
-RU_SCLN, /* ; */
-RU_5,    /* 5 */
-RU_PERC, /* % */
-RU_6,    /* 6 */
-RU_COLN, /* : */
-RU_7,    /* 7 */
-RU_QUES, /* ? */
-RU_8,    /* 8 */
-RU_ASTR, /* * */
-RU_9,    /* 9 */
-RU_LPRN, /* ( */
-RU_0,    /* 0 */
-RU_RPRN, /* ) */
-RU_MINS, /* - */
-RU_UNDS, /* _ */
-RU_EQL,  /* = */
-RU_PLUS, /* + */
-
-RU_J,    /* й */
-RU_S_J,  /* Й */
-RU_TS,   /* ц */
-RU_S_TS, /* Ц */
-RU_U,    /* у */
-RU_S_U,  /* У */
-RU_K,    /* к */
-RU_S_K,  /* К */
-RU_JE,   /* е */
-RU_S_JE, /* Е */
-RU_N,    /* н */
-RU_S_N,  /* Н */
-RU_G,    /* г */
-RU_S_G,  /* Г */
-RU_SH,   /* ш */
-RU_S_SH, /* Ш */
-RU_SC,   /* щ */
-RU_S_SC, /* Щ */
-RU_Z,    /* з */
-RU_S_Z,  /* З */
-RU_H,    /* х */
-RU_S_H,  /* Х */
-RU_HD,   /* ъ */
-RU_S_HD, /* Ъ */
-
-RU_F,    /* ф */
-RU_S_F,  /* Ф */
-RU_Y,    /* ы */
-RU_S_Y,  /* Ы */
-RU_V,    /* в */
-RU_S_V,  /* В */
-RU_A,    /* а */
-RU_S_A,  /* А */
-RU_P,    /* п */
-RU_S_P,  /* п */
-RU_R,    /* р */
-RU_S_R,  /* Р */
-RU_O,    /* о */
-RU_S_O,  /* О */
-RU_L,    /* л */
-RU_S_L,  /* Л */
-RU_D,    /* д */
-RU_S_D,  /* Д */
-RU_ZH,   /* ж */
-RU_S_ZH, /* Ж */
-RU_E,    /* э */
-RU_S_E,  /* Э */
-RU_BSLS, /* \ */
-RU_SLSH, /* / */
-
-RU_JA,   /* я */
-RU_S_JA, /* Я */
-RU_CH,   /* ч */
-RU_S_CH, /* Ч */
-RU_S,    /* с */
-RU_S_S,  /* С */
-RU_M,    /* м */
-RU_S_M,  /* М */
-RU_I,    /* и */
-RU_S_I,  /* И */
-RU_T,    /* т */
-RU_S_T,  /* Т */
-RU_SF,   /* ь */
-RU_S_SF, /* Ь */
-RU_B,    /* б */
-RU_S_B,  /* Б */
-RU_JU,   /* ю */
-RU_S_JU, /* Ю */
-RU_DOT,  /* . */
-RU_COMM, /* , */
-/* -------------------------------------------------------------------- */
-/* Символы, нажимаемые независимо от текущего языка. */
-AG_1,    /* 1 */
-AG_2,    /* 2 */
-AG_3,    /* 3 */
-AG_4,    /* 4 */
-AG_5,    /* 5 */
-AG_6,    /* 6 */
-AG_7,    /* 7 */
-AG_8,    /* 8 */
-AG_9,    /* 9 */
-AG_0,    /* 0 */
-AG_EXCL, /* ! */
-AG_PERC, /* % */
-AG_ASTR, /* * */
-AG_LPRN, /* ( */
-AG_RPRN, /* ) */
-AG_MINS, /* - */
-AG_UNDS, /* _ */
-AG_EQL,  /* = */
-AG_PLUS, /* + */
-AG_SCLN, /* ; */
-AG_COLN, /* : */
-AG_DQUO, /* " */
-AG_BSLS, /* \ */
-AG_COMM, /* , */
-AG_DOT,  /* . */
-AG_SLSH, /* / */
-AG_QUES, /* ? */
