--- lang_shift/keycodes/lang_shift_processing.h
@@ -1,24 +1,0 @@
-SHF_1,   /* Работает как обычный шифт, при её зажатии, шифт зажат. Заодно переключает на 1 слой. */
-SHF_3,   /* Аналогично предыдущей, но переключает на 3 слой. */
-SHF_1_O, /* Данная клавиша применит шифт и переключение на 1 слой только к следующему символу. Неважно, была ли она нажата одновременно со следующей клавишей, или до неё, или вместе с двумя клавишами. Применится только к одной клавише. То есть это как залипающий шифт, если нажимать отдельно от всех клавиш, и как одиночный шифт, если эту клавишу зажимать. Такая клавиша может пригодиться если вы печатаете с большой скоростью и вам сложно выдерживать время удержания шифта, и у вас получается что-то вроде "ПРивет". */
-SHF_3_O, /* Аналогично предыдущей, но переключает на 3 слой. */
-SHF_N,   /* Аналогично SHF_1, но перенаправляет на следующий слой относительно текущего языка (если язык 0, то включает слой 1, если язык 1, то включает слой 3). */
-SHF_N_O, /* Аналогично SHF_1_0, но включает следующий слой относительно текущего языка. */
-LA_CHNG, /* Переключает язык и переключает на 2 или 0 слой в зависимости от текущего языка. */
-LA_SYNC, /* Кнопка для синхронизации языков. Её нужно нажимать в случае если язык в системе и язык в клавиатуре различаются. Данная клавиша изменит язык в системе. */
-LA_CAPS, /* Задаёт переключение языка на Caps. */
-LA_ALSH, /* Задаёт переключение языка на Alt + Shift. */
-LA_CTSH, /* Задаёт переключение языка на Ctrl + Shift. */
-LA_WISP, /* Задаёт переключение языка на Win + Shift. */
-
-AG_3DOT, /* ... */
-AG_SDOT, /* Space + Dot + Shift to next key */
-AG_CMSP, /* Comma + Space */
-
-CTRL_0, /* Ctrl on 0 layer */
-ALT_0,  /* Alt on 0 layer */
-WIN_0,  /* Win on 0 layer */
-SHAL_0, /* Shift+Alt on 0 layer */
-CTAL_0, /* Ctrl+Alt on 0 layer */
-CTSH_0, /* Ctrl+Shift on 0 layer */
-MCAS_0, /* Ctrl+Alt+Shift on 0 layer */
