--- lang_shift/keycodes/shift.h
@@ -1,97 +1,0 @@
-KS_GRV,  /* ` */
-KS_TILD, /* ~ */
-KS_1,    /* 1 */
-KS_EXCL, /* ! */
-KS_2,    /* 2 */
-KS_AT,   /* @ */
-KS_3,    /* 3 */
-KS_HASH, /* # */
-KS_4,    /* 4 */
-KS_DLR,  /* $ */
-KS_5,    /* 5 */
-KS_PERC, /* % */
-KS_6,    /* 6 */
-KS_CIRC, /* ^ */
-KS_7,    /* 7 */
-KS_AMPR, /* & */
-KS_8,    /* 8 */
-KS_ASTR, /* * */
-KS_9,    /* 9 */
-KS_LPRN, /* ( */
-KS_0,    /* 0 */
-KS_RPRN, /* ) */
-KS_MINS, /* - */
-KS_UNDS, /* _ */
-KS_EQL,  /* = */
-KS_PLUS, /* + */
-
-KS_Q,    /* q */
-KS_S_Q,  /* Q */
-KS_W,    /* w */
-KS_S_W,  /* W */
-KS_E,    /* e */
-KS_S_E,  /* E */
-KS_R,    /* r */
-KS_S_R,  /* R */
-KS_T,    /* t */
-KS_S_T,  /* T */
-KS_Y,    /* y */
-KS_S_Y,  /* Y */
-KS_U,    /* u */
-KS_S_U,  /* U */
-KS_I,    /* i */
-KS_S_I,  /* I */
-KS_O,    /* o */
-KS_S_O,  /* O */
-KS_P,    /* p */
-KS_S_P,  /* P */
-KS_LBRC, /* [ */
-KS_LCBR, /* { */
-KS_RBRC, /* ] */
-KS_RCBR, /* } */
-
-KS_A,    /* a */
-KS_S_A,  /* A */
-KS_S,    /* s */
-KS_S_S,  /* S */
-KS_D,    /* d */
-KS_S_D,  /* D */
-KS_F,    /* f */
-KS_S_F,  /* F */
-KS_G,    /* g */
-KS_S_G,  /* G */
-KS_H,    /* h */
-KS_S_H,  /* H */
-KS_J,    /* j */
-KS_S_J,  /* J */
-KS_K,    /* k */
-KS_S_K,  /* K */
-KS_L,    /* l */
-KS_S_L,  /* L */
-KS_SCLN, /* ; */
-KS_COLN, /* : */
-KS_QUOT, /* ' */
-KS_DQUO, /* " */
-KS_BSLS, /* \ */
-KS_PIPE, /* | */
-
-KS_Z,    /* z */
-KS_S_Z,  /* Z */
-KS_X,    /* x */
-KS_S_X,  /* X */
-KS_C,    /* c */
-KS_S_C,  /* C */
-KS_V,    /* v */
-KS_S_V,  /* V */
-KS_B,    /* b */
-KS_S_B,  /* B */
-KS_N,    /* n */
-KS_S_N,  /* N */
-KS_M,    /* m */
-KS_S_M,  /* M */
-KS_COMM, /* , */
-KS_LT,   /* < */
-KS_DOT,  /* . */
-KS_GT,   /* > */
-KS_SLSH, /* / */
-KS_QUES, /* ? */
